+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 * [14.04.2023]
   * neue Anleitungen (viele Modbrix sind dabei)
-  * neue Produktreihe Brix wird mit aufgenommen  
+  * neue Produktreihe BRIX wird mit aufgenommen  
   * Neues Tag Brickbuddies
-  * TODO: "105435" maken history, find new Brix products
 
 * [26.02.2023]
   * neue Anleitungen